Remove table header links - no more sorting
The sort function is currently broken because we sort using CW first
and then take the first "top"(10?) results and sort them according to
&sort=foo. Instead we should only sort according to &sort=foo. This
would involve changes to tor-relay-status.py, so shelving this function
for now. I'm only removing the table header links, and leaving the
sorting code as such so that we can use it when we try to fix it.